back-end/pyproject.toml
[project]
name = "sg-lib-books"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.4",
    "google-cloud-tasks>=2.18.0",
    "nlb-catalogue-client>=0.1.3",
    "pydantic-settings>=2.6.1",
    "supabase>=2.10.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "ruff>=0.7.3",
]

[tool.pyright]
venvPath="."
venv=".venv"
